General use macro: Switch back to frontmost app
This is a macro that saves the frontmost app to a variable so that you
can display a user prompt or let the user do something else before
Keyboard Maestro inserts whatever the user has done back into the
previous frontmost app. I'm making use of this in my Markdown macros.